+++ src/DNET/TCP/TCPServer.h
@@ -104,7 +104,7 @@ class TCPServer
      *
      * @returns 发送成功的数据长度.
      */
-    int Send(int tcpID, const char* data, int len);
+    int Send(int tcpID, const char* data, size_t len);
 
     /**
      * 可读取(接收)的数据数.
